Implemented links on locations and phone numbers
Implemented the ability to click either the individual location or the phone number. Regardless of device, clicking the phone number will prompt the user to actually call that number (pc would prompt an app like Skype or similar). Clicking the location simply brings you to that location on google maps.