ROUTING FIX
this is how we should be routing to different pages within the app going forward.

It is not good to use the href param, we should instead be importing Link from react-router-dom and then using the component and to params (throughout all routing within the app)

For reference, routing is working for now, but this is something that could cause issues in the future if not handled now.